+++ src/Tests/Framework/VirtualClustering/VirtualizedCluster.cs
@@ -12,12 +12,12 @@ namespace Tests.Framework
 		private ElasticClient _client;
 		private readonly VirtualCluster _cluster;
 		private readonly IConnectionPool _connectionPool;
-		private readonly TestDateTimeProvider _dateTimeProvider;
+		private readonly TestableDateTimeProvider _dateTimeProvider;
 		private FixedPipelineFactory _fixedRequestPipeline;
 
 		public VirtualizedCluster(VirtualCluster cluster, IConnectionPool pool, ConnectionSettings settings)
 		{
-			this._dateTimeProvider = new TestDateTimeProvider();
+			this._dateTimeProvider = new TestableDateTimeProvider();
 			this._fixedRequestPipeline = new FixedPipelineFactory(settings, this._dateTimeProvider);
 			this._client = this._fixedRequestPipeline.Client;
 
@@ -30,13 +30,6 @@ namespace Tests.Framework
 		public async Task<ISearchResponse<Project>> ClientCallAsync() => await this._client.SearchAsync<Project>(s => s);
 
 		public void ChangeTime(Func<DateTime, DateTime> change) => change(_dateTimeProvider.MutableNow);
-	}
-
-	public class TestDateTimeProvider : DateTimeProvider
-	{
-		public DateTime MutableNow { get; set; } = DateTime.UtcNow;
-
-		public override DateTime Now() => MutableNow;
 	}
 
 }
