+++ src/Tests/Tests.csproj
@@ -80,15 +80,18 @@
     <Compile Include="ClientConcepts\ConnectionPooling\BuildingBlocks\ConnectionPooling.cs" />
     <Compile Include="ClientConcepts\ConnectionPooling\BuildingBlocks\RequestPipeline.cs" />
     <Compile Include="Framework\Audit\Audits.cs" />
+    <Compile Include="Framework\Audit\AuditTrailTester.cs" />
     <Compile Include="Framework\Audit\ClientCall.cs" />
     <Compile Include="Framework\Audit\ClientCallAssertations.cs" />
     <Compile Include="ClientConcepts\ConnectionPooling\Sniffing\OnStartup.cs" />
     <Compile Include="Framework\Extensions\Numeric.cs" />
+    <Compile Include="Framework\Providers\TestableDateTimeProvider.cs" />
     <Compile Include="Framework\VirtualClustering\Cluster.cs" />
     <Compile Include="Framework\VirtualClustering\FixedPipelineFactory.cs" />
-    <Compile Include="Framework\VirtualClustering\ISniffRule.cs" />
+    <Compile Include="Framework\VirtualClustering\Rules.cs" />
     <Compile Include="Framework\VirtualClustering\SealedVirtualCluster.cs" />
     <Compile Include="Framework\VirtualClustering\VirtualCluster.cs" />
+    <Compile Include="Framework\VirtualClustering\VirtualClusterConnection.cs" />
     <Compile Include="Framework\VirtualClustering\VirtualizedCluster.cs" />
     <Compile Include="ClientConcepts\HighLevel\Inferrence\FieldNames\FieldInference.cs" />
     <Compile Include="ClientConcepts\LowLevel\ConnectionPools\SslDetection.cs" />
